# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RIDB Normalizer API",
    "version": "1.0.0",
    "description": "API documentation for RIDB normalization endpoints used for Campvue and AI integrations."
  },
  "servers": [
    {
      "url": "https://ridb-normalizer.vercel.app",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local dev server"
    }
  ],
  "tags": [
    {
      "name": "Activities",
      "description": "Activities are normalized from the RIDB /activities and /facilities/{id}/activities endpoints."
    },
    {
      "name": "Facilities",
      "description": "Facilities and related metadata normalized from the RIDB /facilities endpoints."
    }
  ],
  "paths": {
    "/api/v1/ridb/activities": {
      "get": {
        "tags": [
          "Activities"
        ],
        "summary": "Fetch and normalize RIDB activities data.",
        "description": "Returns a normalized view of activities from the Recreation Information Database (RIDB).\nBy default (when `getAll` is not provided or false), this endpoint returns a **single page** of activities, controlled by the `limit` and `offset` parameters.\nWhen `getAll` is true, the endpoint will page through the RIDB /activities API using a page size of 50 and return **all** activities in a single response. In this mode, any `limit` and `offset` query parameters are ignored, even if they appear in the URL (including Swagger UI defaults).\n",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "description": "Query filter criteria. Searches on activity name using RIDB's substring matcher (e.g. `query=biking` will match \"Biking\"). If left blank, this parameter is ignored and all activities are returned.\n",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "description": "Page size for a single request to RIDB. Must be an integer between 1 and 50 (RIDB maximum). Defaults to **50** when not provided. Ignored when `getAll=true`.\n",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "example": 50,
              "default": 50
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "description": "Start record of the overall result set, as defined by the RIDB API. Must be a non-negative integer. Defaults to **0** when not provided. Ignored when `getAll=true`.\n",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "example": 0,
              "default": 0
            }
          },
          {
            "in": "query",
            "name": "getAll",
            "required": false,
            "description": "When true, this endpoint will page through the RIDB /activities API and return **all** activities in a single response. In this mode, the `limit` and `offset` query parameters are ignored, even if provided (including when Swagger UI shows default values of 50 and 0).\n",
            "schema": {
              "type": "boolean",
              "example": false,
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Normalized activities"
          },
          "400": {
            "description": "Invalid query parameters (e.g. limit or offset)"
          },
          "500": {
            "description": "RIDB error or server error"
          }
        }
      }
    },
    "/api/v1/ridb/activitiesByFacilityId": {
      "get": {
        "tags": [
          "Activities"
        ],
        "summary": "Fetch and normalize RIDB facility activitiesByFacilityId data.",
        "description": "Returns a normalized view of a single facility's activities by facility id from the Recreation Information Database (RIDB). You can try this using a facility ID taken from a recreation.gov campground URL.\nFor example, the Mather Campground URL: https://www.recreation.gov/camping/campgrounds/232490 has facility ID **232490**.\nThis endpoint will page through RIDB results using the provided `limit` and `offset` values (or their defaults) and return all activities for the facility starting at that offset. You can optionally use `take` to limit the number of normalized items returned by this endpoint.\n",
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "description": "RIDB FacilityID (numeric ID from the recreation.gov campground URL), e.g. **232490** for Mather Campground at Grand Canyon.\n",
            "schema": {
              "type": "string",
              "example": "232490",
              "default": "232490"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "description": "Number of activity records to request from RIDB for this facility per page (page size). Must be an integer between 1 and 50 (RIDB maximum). Defaults to **50** when not provided.\n",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "example": 50,
              "default": 50
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "description": "Start record of the overall result set, as defined by the RIDB API. Must be a non-negative integer. Defaults to **0** when not provided.\n",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "example": 0,
              "default": 0
            }
          },
          {
            "in": "query",
            "name": "take",
            "required": false,
            "description": "Maximum number of normalized activity items to return **from this endpoint**. Unlike RIDB's `limit` (page size), `take` limits the final response array after all paging and normalization is complete. If omitted or set to `0`, all normalized results are returned. This feature does not exist in the RIDB API.\n",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "example": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Normalized activities by facility"
          },
          "400": {
            "description": "Missing or invalid query parameters (e.g. id, limit, offset, or take)"
          },
          "500": {
            "description": "RIDB error or server error"
          }
        }
      }
    },
    "/api/v1/ridb/activityById": {
      "get": {
        "tags": [
          "Activities"
        ],
        "summary": "Fetch and normalize RIDB a single activity's data.",
        "description": "Returns a normalized view of a single activity by activity id from the Recreation Information Database (RIDB). You need an activity id to use this api, e.g. 4, 50, 75...  ids are easily obtained from the list of all activities here at /activities or via the RIDB endpoint.\nFor example, the RIDB activities endpoint exposes: {\n \"ActivityID\": 4,\n \"ActivityLevel\": 0,\n \"ActivityName\": \"AUTO TOURING\",\n \"ActivityParentID\": 0\n} and has activity ID **4**.\nThis endpoint returns one activity by id at a time. No paging required.  If an activity by id is not found, note the 404 spec below.\n",
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "description": "RIDB ActivityID (numeric ID from the recreation.gov list of activities), e.g. **4** for Auto Touring.\n",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Normalized activities by activity id"
          },
          "400": {
            "description": "Missing id"
          },
          "404": {
            "description": "Activity not found (RIDB returned id 0 and empty name)"
          },
          "500": {
            "description": "RIDB error or server error"
          }
        }
      }
    },
    "/api/v1/ridb/facility": {
      "get": {
        "tags": [
          "Facilities"
        ],
        "summary": "Fetch and normalize RIDB facility data.",
        "description": "Returns a normalized view of a single facility from the Recreation Information Database (RIDB). You can try this using a facility ID taken from a recreation.gov campground URL.\nFor example, the Mather Campground URL: https://www.recreation.gov/camping/campgrounds/232490 has facility ID **232490**.\n",
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "description": "RIDB FacilityID (numeric ID from the recreation.gov campground URL), e.g. **232490** for Mather Campground at Grand Canyon.\n",
            "schema": {
              "type": "string",
              "example": "232490",
              "default": "232490"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Normalized facility"
          },
          "400": {
            "description": "Missing ID"
          },
          "500": {
            "description": "RIDB error or server error"
          }
        }
      }
    }
  },
  "components": {}
}